extract inventory container controller and inventory item controller
